+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED] - YYYY-MM-DD
 
+### Fixed
+- [#429](https://github.com/equinor/webviz-config/pull/429) - Moved `FutureWarnings`
+from `deprecated_decorators.py` to `_config_parser.py`. Simplified deprecation warnings.
+
 ## [0.3.0] - 2021-04-27
 
 ### Added